feat: controller 구현, api test 용 더미데이터 등록

+++ src/main/java/com/conduit/analytics/dto/repository/HourStatDto.java
@@ -10,7 +10,7 @@ public class HourStatDto {
 
     private TimePeriod timePeriod;  // 시간대 (새벽, 오전, 오후, 야간)
     private String actionType;      // 행동 유형 (VIEW, CLICK, VOTE 등)
-    private Long totalActionCount;  // 해당 시간대의 총 행동 횟수
+    private Integer totalActionCount;  // 해당 시간대의 총 행동 횟수
     private Long uniqueUserCount;   // 해당 시간대에 행동한 고유 사용자 수
